requirements/core_requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==2.2.5
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
pandas
sendgrid
requests
sqlalchemy
gspread
google-auth
openpyxl
sentence-transformers